Fixed clock plugin

+++ plugins/clock/clock.cpp
@@ -6,15 +6,33 @@
 #include <format>
 #include <imgui.h>
 
-using EnqueueFn = void (*)(std::function<void()>&&);
-using CreateCommandFn = void (*)(const wchar_t*, const GW::Chat::CmdCB&);
-using DeleteCommandFn = void (*)(const wchar_t*);
-using SendChatFn = void (*)(char, const char*);
+namespace {
+    bool send_to_chat = false;
 
-EnqueueFn enqueue = nullptr;
-CreateCommandFn create_command = nullptr;
-DeleteCommandFn delete_command = nullptr;
-SendChatFn send_chat = nullptr;
+    using CreateCommandFn = void (*)(const wchar_t*, const GW::Chat::CmdCB&);
+    using DeleteCommandFn = void (*)(const wchar_t*);
+    using SendChatFn = void (*)(char, const char*);
+
+    CreateCommandFn create_command = nullptr;
+    DeleteCommandFn delete_command = nullptr;
+    SendChatFn send_chat = nullptr;
+
+    void CmdClock(const wchar_t*, int, wchar_t**)
+    {
+        send_to_chat = true;
+    }
+    std::string GetTime()
+    {
+        const auto now = std::chrono::system_clock::now();
+        const auto time = std::chrono::system_clock::to_time_t(now);
+        char buf[100];
+        ctime_s(buf, sizeof buf, &time);
+        auto str = std::format("{}", buf);
+        str.pop_back();
+        return str;
+    }
+}
+
 
 DLLAPI ToolboxPlugin* ToolboxPluginInstance()
 {
@@ -22,31 +40,23 @@ DLLAPI ToolboxPlugin* ToolboxPluginInstance()
     return &instance;
 }
 
-auto GetTime()
-{
-    const auto now = std::chrono::system_clock::now();
-    const auto time = std::chrono::system_clock::to_time_t(now);
-    char buf[100];
-    ctime_s(buf, sizeof buf, &time);
-    auto str = std::format("{}", buf);
-    str.pop_back();
-    return str;
-}
 
+void Clock::Update(float)
+{
+    if (!toolbox_handle)
+        return;
+    if (send_to_chat && send_chat) {
+        send_chat('#', GetTime().c_str());
+        send_to_chat = false;
+    }
+}
 void Clock::Draw(IDirect3DDevice9*)
 {
     if (!toolbox_handle)
         return;
     ImGui::Begin("clock");
-    ImGui::Text("%s", GetTime().c_str());
+    ImGui::TextUnformatted(GetTime().c_str());
     ImGui::End();
-}
-
-void CmdClock(const wchar_t*, int, wchar_t**)
-{
-    enqueue([] {
-        send_chat('#', GetTime().c_str());
-    });
 }
 
 void Clock::Initialize(ImGuiContext* ctx, ImGuiAllocFns fns, HMODULE toolbox_dll)
@@ -54,16 +64,16 @@ void Clock::Initialize(ImGuiContext* ctx, ImGuiAllocFns fns, HMODULE toolbox_dll
     ToolboxPlugin::Initialize(ctx, fns, toolbox_dll);
 
     // we load our gwca methods dynamically
-    enqueue = reinterpret_cast<EnqueueFn>(GetProcAddress(toolbox_dll, "?Enqueue@GameThread@GW@@YAXV?$function@$$A6AXXZ@std@@@Z"));
     create_command = reinterpret_cast<CreateCommandFn>(GetProcAddress(toolbox_dll, "?CreateCommand@Chat@GW@@YAXPB_WABV?$function@$$A6AXPB_WHPAPA_W@Z@std@@@Z"));
     delete_command = reinterpret_cast<DeleteCommandFn>(GetProcAddress(toolbox_dll, "?DeleteCommand@Chat@GW@@YAXPB_W@Z"));
     send_chat = reinterpret_cast<SendChatFn>(GetProcAddress(toolbox_dll, "?SendChat@Chat@GW@@YAXDPBD@Z"));
-
-    create_command(L"clock", CmdClock);
+    if(create_command)
+        create_command(L"clock", CmdClock);
 }
 
 void Clock::Terminate()
 {
     ToolboxPlugin::Terminate();
-    delete_command(L"clock");
+    if(delete_command)
+        delete_command(L"clock");
 }
